fix: int32 offset overflow in closure and block-sparse conv triton kernels (Bug 1)
Triton 3.5.1 wraps int32 address arithmetic silently rather than trapping.
At hourly resolution (768 taps) two independent products exceed 2^31:

- conv kernels (fwd, bwd_dx, bwd_dvalues): nzb*(K*M*M) weight-block offset
  wraps past ~10,922 nonzero blocks; the batch/block time-strides
  (n_blocks*T*M) separately wrap once T*block_count crosses the same
  threshold. Confirmed via a direct triton wrap probe and by cross-checking
  a wrapping deep-path batch (18,584 blocks) against the identical reaches
  requested in a row order that stays under the threshold (5,469 blocks) --
  both now agree to floating-point reordering noise (rel 2.6e-7).
- closure kernels (_coo_enum_sum_kernel, _vals_to_prefix_grad_kernel):
  row*n_feat wraps once path_cumsum[-1]*n_feat exceeds 2^31 (the full
  45,338-output VPU 209 hourly closure: 6,089,353 paths x 770 features =
  2.18x2^31). Node-indexed products (pid/dest/start * n_feat) are left
  alone -- they stay under 2^31 for any graph size this library routes.

Fix: widen the first large factor of each offset product to int64 before
multiplying; downstream arithmetic promotes through it. No value, dtype,
or floating-point computation order changes -- this is address arithmetic
only. Bitwise regression check (scripts/d_certify_bitwise.py in camels_tdx)
against pre-fix references: case A (daily clustered) forward bitwise
identical, all other diffs at the pre-existing atomics run-to-run noise
floor, not a new discrepancy.

Also, while here:
- kernels.py: from_coo uses index_put_ (in place) instead of index_put --
  linear_indices are unique by construction (coords has no duplicate
  (dest,src) rows), so there is nothing to accumulate, and this drops a
  same-size clone of block_values from peak memory.
- prefix_sum.py: PrefixSum.backward passes max_depth=n to skip
  _prefix_jump_bwd's default O(n . depth) CPU chase over `edges` (a python
  loop with a device sync) every backward call. n is a trivially safe
  upper bound; the round loop already breaks early once no more gradient
  mass moves, so no additional rounds actually execute -- this only
  removes dead CPU work, independently valuable at the 1.84M-reach global
  graph scale (STATE.md SS5).

Does not yet address the representation cost of large-K block-sparse
storage itself (Task D "beyond a certain kernel window", part b) -- a
4096-output hourly batch still OOMs at 80GB on backward even fixed
(42.7k blocks at 2-4% fill). That is the next phase.

+++ diffroute/ops/closure_sub.py
@@ -67,7 +67,12 @@ def _coo_enum_sum_kernel(coords_ptr, vals_ptr,
         # emitted kernel is (n_out x n_in) rather than (n x n) with holes.
         out_row = tl.load(out_row_ptr + dest)
         if out_row >= 0:
-            row = base + step
+            # row indexes PATHS (up to path_cumsum[-1]), which for a large
+            # closure times n_feat exceeds int32 -- widen before the store
+            # offset is computed, or the product silently wraps (Triton does
+            # not trap on overflow). pid/dest index NODES, which stay well
+            # under 2^31 even at n_feat*n_nodes for the largest graphs routed.
+            row = (base + step).to(tl.int64)
 
             tl.store(coords_ptr + row*2 + 0, out_row)
             tl.store(coords_ptr + row*2 + 1, pid)
@@ -89,7 +94,7 @@ def _vals_to_prefix_grad_kernel(coords_ptr, out_pos_ptr, gvals_ptr,
                                 ghead_ptr, gtail_ptr,
                                 n_feat: tl.constexpr,
                                 BLOCK_F: tl.constexpr):
-    row  = tl.program_id(0)
+    row  = tl.program_id(0).to(tl.int64)  # indexes PATHS; see the forward's note
     offs = tl.arange(0, BLOCK_F)
 
     # coords[:, 0] is the destination's row in the OUTPUT (see the forward's
